7_Bedrock_Agents/frontend/pyproject.toml
[project]
name = "bedrock-agents-frontend"
version = "0.1.0"
description = "Streamlit app for the AI Agent Insure Bedrock Agent — calls Bedrock directly via boto3"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.35.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]
